refactor: rename ava helpers file (> plural)
